+++ src/nix/health/mod.rs
@@ -1,19 +1,26 @@
 //! Health checks for the user's Nix install
 
+mod max_jobs;
+
 use leptos::*;
 use serde::{Deserialize, Serialize};
 use tracing::instrument;
 
+use self::max_jobs::MaxJobs;
 use super::info;
 
+/// Nix Health check information
+///
+/// This struct is isomorphic to Vec<Box<&dyn Check>>. We cannot use the latter
+/// due to (wasm) serialization limitation with dyn trait objects.
 #[derive(Debug, PartialEq, PartialOrd, Eq, Ord, Serialize, Deserialize, Clone)]
-pub struct NixHealthChecks {
+pub struct NixHealth {
     max_jobs: MaxJobs,
 }
 
-impl NixHealthChecks {
+impl NixHealth {
     pub fn new(info: info::NixInfo) -> Self {
-        NixHealthChecks {
+        NixHealth {
             max_jobs: MaxJobs::check(info),
         }
     }
@@ -39,35 +46,17 @@ pub trait Check {
     fn report(&self) -> Report;
 }
 
-#[derive(Debug, PartialEq, PartialOrd, Eq, Ord, Serialize, Deserialize, Clone)]
-pub struct MaxJobs(i32);
-
-impl Check for MaxJobs {
-    fn check(info: info::NixInfo) -> Self {
-        MaxJobs(info.nix_config.max_jobs.value)
-    }
-    fn report(&self) -> Report {
-        if self > &MaxJobs(1) {
-            Report::Green
-        } else {
-            Report::Red {
-                msg: "You are using only 1 core for nix builds",
-                suggestion: "Try editing /etc/nix/nix.conf",
-            }
-        }
-    }
-}
-
 #[instrument(name = "nix-health")]
 #[server(GetHealthChecks, "/api")]
-pub async fn get_health_checks() -> Result<NixHealthChecks, ServerFnError> {
+pub async fn get_health_checks() -> Result<NixHealth, ServerFnError> {
     let info = info::get_nix_info().await?;
-    Ok(NixHealthChecks::new(info))
+    Ok(NixHealth::new(info))
 }
 
-impl IntoView for NixHealthChecks {
+impl IntoView for NixHealth {
     fn into_view(self, cx: Scope) -> View {
         let checks = self.as_list();
+        // TODO: list
         let check0 = checks[0].report();
         view! { cx, <pre>{format!("{:?}", check0)}</pre> }.into_view(cx)
     }
